factored out functions and vars that both pr2 and fetch had in common
